added frontend module and extended backend

+++ backend/src/main/kotlin/com/ephemeris/backend/CardRepository.kt
@@ -1,9 +1,10 @@
 package com.ephemeris.backend
 
 import org.springframework.data.mongodb.repository.MongoRepository
+import java.math.BigInteger
 import java.util.*
 
-interface CardRepository : MongoRepository<Card, Long> {
+interface CardRepository : MongoRepository<Card, BigInteger> {
 
     fun findByDate(date: Date): List<Card>
 }
